Updated Dao to have Long as findById argument

+++ deploy-core/src/main/java/uk/laxd/deepweb/dao/Dao.java
@@ -6,6 +6,6 @@ import java.util.List;
  * Created by Lenny on 27/09/2015.
  */
 public interface Dao<T> {
-    T findById(int id);
+    T findById(Long id);
     List<T> findByRestrictions(Restrictions restriction);
 }
